refactor: core codes

+++ src/core/create-react-component.ts
@@ -1,54 +1,70 @@
 import * as vscode from 'vscode';
 import * as fs from 'fs-extra';
-import * as htmlTemplates from '../templates/html-elements/div';
+import * as _htmlTemplates from '../templates/html-elements';
+
+
+type UsedElement = keyof typeof _htmlTemplates;
+
+const htmlTemplates = _htmlTemplates;
 
 export async function createReactComponentAction() {
-  const input = await vscode.window.showInputBox({
-    title: 'Please Input the component name',
-    placeHolder: 'UserNameInput',
-  });
-
-  const elInput = await vscode.window.showInputBox({
-    title: 'Please HTML Element name',
-    placeHolder: 'div, img',
-  });
+  
+  const { componentNameInput, htmlElementInput } = await getUserInputs();
 
   // if Not input
-  if (!input || !elInput || !Object.keys(htmlTemplates).includes(elInput)) {
+  if (
+    !componentNameInput ||
+    !htmlElementInput ||
+    !Object.keys(htmlTemplates).includes(htmlElementInput)
+  ) {
     vscode.window.showErrorMessage('Please check the input');
     return;
   }
 
-  // create new file
+
+  /**
+   * Create new files
+   */
+
   const wsEdit = new vscode.WorkspaceEdit();
   const wsPath = vscode.workspace.workspaceFolders![0].uri.fsPath;
+
   // TODO: Add a method of modify the component path
   const componentFilePath = vscode.Uri.file(
-    `${wsPath}/components/${input}/${input}.tsx`
+    wsPath + `/components/${componentNameInput}/${componentNameInput}.tsx`
   );
+
   const fileIndexPath = vscode.Uri.file(
-    `${wsPath}/components/${input}/index.ts`
+    wsPath + `/components/${componentNameInput}/index.ts`
   );
+
+  if (fs.existsSync(componentFilePath.fsPath) || fs.existsSync(fileIndexPath.fsPath)) {
+    vscode.window.showErrorMessage('The Component already exists.');
+    return;
+  }
 
   wsEdit.createFile(componentFilePath, { ignoreIfExists: true });
   wsEdit.createFile(fileIndexPath, { ignoreIfExists: true });
   await vscode.workspace.applyEdit(wsEdit);
 
-  // if (fs.existsSync(fileIndexPath.fsPath)) {
-  //   vscode.window.showErrorMessage('The Component already exists.');
-  //   return;
-  // }
+
+  /**
+   * file write
+   */
 
   await fs.writeFile(
     fileIndexPath.fsPath,
-    `export * from './${input}'`,
+    `export * from './${componentNameInput}'`,
     {
       encoding: 'utf8',
     },
-    handleWriteFileError
+    handleWriteFile
   );
 
-  const changedText = htmlTemplates.div.replaceAll(/ComponentName/gi, input);
+  const changedText = htmlTemplates[htmlElementInput].replaceAll(
+    /ComponentName/gi,
+    componentNameInput
+  );
 
   await fs.writeFile(
     componentFilePath.fsPath,
@@ -56,11 +72,26 @@ export async function createReactComponentAction() {
     {
       encoding: 'utf8',
     },
-    handleWriteFileError
+    handleWriteFile
   );
 }
 
-function handleWriteFileError(error: NodeJS.ErrnoException) {
+
+async function getUserInputs(){
+
+  const componentNameInput = await vscode.window.showInputBox({
+    title: 'Please Input the component name',
+    placeHolder: 'UserNameInput',
+  });
+
+  const htmlElementInput = await vscode.window.showInputBox({
+    title: 'Please HTML Element name',
+    placeHolder: 'div, img',
+  });
+
+  return { componentNameInput, htmlElementInput: (htmlElementInput as UsedElement | undefined) }
+}
+function handleWriteFile(error: NodeJS.ErrnoException) {
   if (error) {
     vscode.window.showErrorMessage(error.message);
   } else {
